src/requirements.txt
jsonformatter==0.3.4
prometheus_client==0.22.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
requests==2.32.5
schedule==1.2.2
urllib3==2.5.0
